backend/requirements.txt
fastapi
uvicorn
pymupdf
python-docx
python-multipart
python-dotenv
google-generativeai
scikit-learn
